Move UITableViewDelegate from LanguagesManager to its own file extending the LanguageSettingsTableViewController
Also:
 - set languages in the view controller
 - hold a reference to an instance of languages manager, not a static shared
 - make a couple methods public now that things out side the languages manager call them
